Burn Effect (On Baba)

+++ NO2/Assets/Scripts/World/Enemies/IEffectable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IEffectable
 {
-    void ApplyBurn(float damagePerSecond, float duration);
+    void ApplyBurn(float damagePerSecond, int duration);
 }
